requirements.txt
Django==4.1.1
djangorestframework==3.14.0
psycopg2==2.9.3
pillow==9.2.0
django-cors-headers==3.13.0
django-colorfield==0.4.1
whitenoise==6.2.0
gunicorn==20.1.0
dj-database-url==0.5.0
django-storages==1.11.1
boto3
django-debug-toolbar==3.6.0
django-jazzmin==2.5.0
celery==5.2.7
redis
xlrd

# auth
dj-rest-auth==2.2.5
django-allauth==0.43.0
djangorestframework-simplejwt

# phone
pyotp==2.7.0
twilio==7.14.2